Extract the action creators and middleware.
Getting the easy parts done first.
